+++ CHANGES.rst
@@ -9,6 +9,8 @@ WIP
 - `organization owners <https://github.com/openwisp/openwisp-users#organization-owners>`_:
   [admin] Added support for organization owners
 - [admin] Made organization owners read-only to non-superusers
+- `organization permissions <https://github.com/openwisp/openwisp-users#organization-permissions>`_:
+  [admin] Allowed administrator role to access organization admin
 
 Version 0.2.2 [2020-05-04]
 --------------------------